+++ src/main/java/blackjack/domain/cards/Hand.java
@@ -1,6 +1,5 @@
-package blackjack.domain;
+package blackjack.domain.cards;
 
-import blackjack.domain.card.Card;
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
