+++ src/test/java/blackjack/domain/cards/CardTest.java
@@ -1,4 +1,4 @@
-package blackjack.domain.card;
+package blackjack.domain.cards;
 
 import static org.assertj.core.api.Assertions.assertThat;
