+++ src/test/java/blackjack/domain/participants/PlayersTest.java
@@ -1,14 +1,14 @@
-package blackjack.domain;
+package blackjack.domain.participants;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatIllegalStateException;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
-import blackjack.domain.card.Card;
-import blackjack.domain.card.CardValue;
-import blackjack.domain.card.Shape;
-import blackjack.domain.participant.Dealer;
-import blackjack.domain.participant.Player;
+import blackjack.domain.ResultType;
+import blackjack.domain.cards.Card;
+import blackjack.domain.cards.CardValue;
+import blackjack.domain.cards.Deck;
+import blackjack.domain.cards.Shape;
 import blackjack.dto.GameResult;
 import java.util.Arrays;
 import java.util.HashMap;
